@@ -1,13 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "23.0.1"
+    compileSdkVersion 31
+    buildToolsVersion "29.0.2"
 
     defaultConfig {
         applicationId "com.alamkanak.weekview"
         minSdkVersion 9
-        targetSdkVersion 23
+        targetSdkVersion 31
         versionCode 1
         versionName "1.0"
     }
@@ -20,8 +20,9 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile project(':library')
-    compile 'com.android.support:appcompat-v7:23.1.1'
-    compile 'com.squareup.retrofit:retrofit:1.9.0'
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation project(':library')
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation("androidx.interpolator:interpolator:1.0.0")
+    implementation 'com.squareup.retrofit:retrofit:1.9.0'
 }
